LPS-83992 PortletScannerUtil has nothing to do with xml, move to util package
